refactor(bin): extract createTimer function
